+++ eCommerce/src/main/java/com/apiFinal/eCommerce/services/ItemPedidoService.java
@@ -6,10 +6,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.apiFinal.eCommerce.entities.ItemPedido;
+import com.apiFinal.eCommerce.exceptions.UniqueElementException;
+import com.apiFinal.eCommerce.exceptions.UnmatchingIdsException;
 import com.apiFinal.eCommerce.repositories.ItemPedidoRepository;
 
 @Service
 public class ItemPedidoService {
+
 
 	@Autowired
 	ItemPedidoRepository itemPedidoRepository;
@@ -23,20 +26,41 @@ public class ItemPedidoService {
 	}
 	
 	public ItemPedido saveItemPedido(ItemPedido itemPedido) {
-		return itemPedidoRepository.save(itemPedido);
+		if (itemPedido.getIdItemPedido() == null) {
+			try {
+				return itemPedidoRepository.save(itemPedido);		
+			} catch (Exception e) {
+				throw new UniqueElementException();
+			}			
+		} else {
+			throw new UnmatchingIdsException(itemPedido.getIdItemPedido(), itemPedido.toString());
+		}
 	}
 	
 	public ItemPedido updateItemPedido(ItemPedido itemPedido) {
-		return itemPedidoRepository.save(itemPedido);
+		Integer id = itemPedido.getIdItemPedido();
+		if (id == null) {
+			throw new UnmatchingIdsException();
+		} else {
+			if (itemPedidoRepository.findById(id).orElse(null) != null) {
+				try {
+					return itemPedidoRepository.save(itemPedido);		
+				} catch (Exception e) {
+					throw new UniqueElementException();
+				}					
+			} else {
+				throw new UnmatchingIdsException(id);
+			}
+		}
 	}
 	
 	public Boolean deleteItemPedido(Integer id) {
-		itemPedidoRepository.deleteById(id);
-		ItemPedido itemPedidoDeletado = itemPedidoRepository.findById(id).orElse(null);
-		if(itemPedidoDeletado == null) {
+		ItemPedido itemPedidoDeletada = itemPedidoRepository.findById(id).orElse(null);
+		if(itemPedidoDeletada != null) {
+			itemPedidoRepository.deleteById(id);
 			return true;
 		} else {
-			return false;
+			throw new UnmatchingIdsException(id);
 		}
 	}
