+++ eCommerce/src/main/java/com/apiFinal/eCommerce/services/ProdutoService.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 
 import com.apiFinal.eCommerce.entities.Produto;
 import com.apiFinal.eCommerce.exceptions.UniqueElementException;
+import com.apiFinal.eCommerce.exceptions.UnmatchingIdsException;
 import com.apiFinal.eCommerce.repositories.ProdutoRepository;
 
 @Service
@@ -14,35 +15,52 @@ public class ProdutoService {
 
 	@Autowired
 	ProdutoRepository produtoRepository;
-	
+
 	public List<Produto> getAllProdutos() {
 		return produtoRepository.findAll();
 	}
-	
+
 	public Produto getProdutoById(Integer id) {
 		return produtoRepository.findById(id).orElse(null);
 	}
-	
+
 	public Produto saveProduto(Produto produto) {
-		try {
-			return produtoRepository.save(produto);		
-		} catch (Exception e) {
-			throw new UniqueElementException();
+		if (produto.getIdProduto() == null) {
+			try {
+				return produtoRepository.save(produto);
+			} catch (Exception e) {
+				throw new UniqueElementException();
+			}
+		} else {
+			throw new UnmatchingIdsException(produto.getIdProduto(), produto.getNome());
 		}
 	}
-	
+
 	public Produto updateProduto(Produto produto) {
-		return produtoRepository.save(produto);
+		Integer id = produto.getIdProduto();
+		if (id == null) {
+			throw new UnmatchingIdsException();
+		} else {
+			if (produtoRepository.findById(id).orElse(null) != null) {
+				try {
+					return produtoRepository.save(produto);
+				} catch (Exception e) {
+					throw new UniqueElementException();
+				}
+			} else {
+				throw new UnmatchingIdsException(id);
+			}
+		}
 	}
-	
+
 	public Boolean deleteProduto(Integer id) {
-		produtoRepository.deleteById(id);
-		Produto produtoDeletado = produtoRepository.findById(id).orElse(null);
-		if(produtoDeletado == null) {
+		Produto produtoDeletada = produtoRepository.findById(id).orElse(null);
+		if (produtoDeletada != null) {
+			produtoRepository.deleteById(id);
 			return true;
 		} else {
-			return false;
+			throw new UnmatchingIdsException(id);
 		}
 	}
-	
+
 }
